+++ examples/compute/instance/instance.tf
@@ -140,6 +140,14 @@ resource "oci_core_instance" "test_instance" {
   freeform_tags = {
     "freeformkey${count.index}" = "freeformvalue${count.index}"
   }
+
+  preemptible_instance_config {
+    preemption_action {
+      type = "TERMINATE"
+      preserve_boot_volume = false
+    }
+  }
+
   timeouts {
     create = "60m"
   }