+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Support for VCN API: include OCID of Local Peering Gateway in return model
 - Support for Streaming analytics in OCI Connector Hub
 - Support for IPv6
+- Support for Compute preemptible Instances
 
 ## 4.21.0 (April 07, 2021)